project reconstruction accomplished

+++ src/include/impl/conditional_euclidean_clustering.hpp
@@ -116,7 +116,7 @@ pcl::ConditionalEuclideanClustering<PointT>::segment (pcl::IndicesClusters &clus
     }
 
     // If extracting removed clusters, all clusters need to be saved, otherwise only the ones within the given cluster size range
-    if (extract_removed_clusters_ || (current_cluster.size () >= min_cluster_size_ && current_cluster.size () <= max_cluster_size_))
+    if (extract_removed_clusters_ || ((int)current_cluster.size () >= min_cluster_size_ && (int)current_cluster.size () <= max_cluster_size_))
     {
       pcl::PointIndices pi;
       pi.header = input_->header;
@@ -124,9 +124,9 @@ pcl::ConditionalEuclideanClustering<PointT>::segment (pcl::IndicesClusters &clus
       for (int ii = 0; ii < static_cast<int> (current_cluster.size ()); ++ii)  // ii = indices iterator
         pi.indices[ii] = current_cluster[ii];
 
-      if (extract_removed_clusters_ && current_cluster.size () < min_cluster_size_)
+      if (extract_removed_clusters_ && (int)current_cluster.size () < min_cluster_size_)
         small_clusters_->push_back (pi);
-      else if (extract_removed_clusters_ && current_cluster.size () > max_cluster_size_)
+      else if (extract_removed_clusters_ && (int)current_cluster.size () > max_cluster_size_)
         large_clusters_->push_back (pi);
       else
         clusters.push_back (pi);
